// Neptune.Web/src/app/shared/generated/model/field-visit-create-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class FieldVisitCreateDto { 
    TreatmentBMPID?: number;
    VisitDate?: string;
    FieldVisitTypeID?: number;
    ContinueExistingInProgress?: boolean | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface FieldVisitCreateDtoForm { 
    TreatmentBMPID?: FormControl<number>;
    VisitDate?: FormControl<string>;
    FieldVisitTypeID?: FormControl<number>;
    ContinueExistingInProgress?: FormControl<boolean>;
}

export class FieldVisitCreateDtoFormControls { 
    public static TreatmentBMPID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static VisitDate = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FieldVisitTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ContinueExistingInProgress = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}